+++ src/index.ts
@@ -1,5 +1,5 @@
 export { Caller } from './caller';
 export { ExtError } from './errorMessages';
-export { extCall, onCallWithCaller, setExtCallDefaultRegion, ExtErrorMessages } from './extCall';
+export { extCall, setExtCallDefaults, ExtErrorMessages } from './extCall';
 export type { obj } from './utils';
 export type { Handler, HandlerF, ExtErrorT } from './types';